refactor: move mainnet default constants into networks
